+++ modules/Users/Preferences.php
@@ -747,7 +747,8 @@ if ( ! $_REQUEST['modfunc'] )
 			( User( 'PROFILE_ID' ) ?
 			"PROFILE_EXCEPTIONS WHERE PROFILE_ID='" . User( 'PROFILE_ID' ) . "'" :
 			"STAFF_EXCEPTIONS WHERE USER_ID='" . User( 'STAFF_ID' ) . "'" ) .
-			" AND MODNAME='Users/User.php&category_id='||cf.CATEGORY_ID)='Y'
+			" AND MODNAME='Users/User.php&category_id='||cf.CATEGORY_ID
+			LIMIT 1)='Y'
 			AND cf.TYPE<>'files'
 			ORDER BY sfc.SORT_ORDER,sfc.TITLE,cf.SORT_ORDER,cf.TITLE";